2026-03-12-mergeguard/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mergeguard"
version = "0.1.0"
description = "Defensive CLI tool to filter out unmergeable AI slop from PRs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "PyYAML>=6.0",
    "GitPython>=3.1.0",
    "openai>=1.0.0"
]

[project.scripts]
mergeguard = "mergeguard.cli:app"
